+++ newlogic_main/models/__init__.py
@@ -5,7 +5,10 @@
 # All Rights Reserved.
 #################################################################################
 
-from . import custom_res_partner
+from . import references
+from . import registrant
 from . import reg_individual
 from . import reg_group
-
+from . import reg_id
+from . import reg_relationship
+from . import group_membership