+++ newlogic_main/CHANGELOG.md
@@ -1,5 +1,9 @@
 Format in CHANGELOG.md
 
+## Version[0.04] - 2022-05-05
+- [FIX] Registrant, individual, group models inherits res.partner
+- [ADD] Group membership, ID type, Registrant ID, Relationship, Registrant Relationship models
+
 ## Version[0.03] - 2022-05-04
 - [FIX] res.partner: create m2o fields to link individual and group models
 - [ADD] Create model for individuals (nl.reg.individual)